backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.3
pydantic==2.5.3
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
python-dotenv==1.0.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
sentence-transformers
chromadb
langchain-text-splitters
langdetect
transformers
accelerate
openai-whisper